Price type changed from Double to Long! account page has been added. Created account controller. Deleted getMapping on /account from HomeController

+++ src/main/java/com/stc21/boot/auction/entity/Lot.java
@@ -52,14 +52,14 @@ public class Lot {
     private City city;
 
     @Column(name = "current_price")
-    private Double currentPrice;
+    private Long currentPrice;
 
     @Column(name = "max_price")
-    private Double maxPrice;
+    private Long maxPrice;
 
     @Column(name = "min_price")
-    private Double minPrice;
+    private Long minPrice;
 
     @Column(name = "step_price")
-    private Double stepPrice;
+    private Long stepPrice;
 }
